requirements.txt
Django==6.0.4
djangorestframework==3.17.1
paho_mqtt==2.1.0
pymodbus==3.5.2
pymongo==4.16.0
Requests==2.33.1
